requirements.txt
boto3==1.40.7
botocore==1.40.7
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
urllib3==1.26.20
